ai-news-radar/requirements.txt
feedparser>=6.0.11
requests>=2.31.0
apscheduler>=3.10.4
python-dateutil>=2.8.2
beautifulsoup4>=4.12.3
